feat: rename Entity\LocalizedTrait
